requirements.txt
# Movie Rental Management System - Python Dependencies

# Database Connection
mysql-connector-python>=8.0.0

# Data Processing and Analysis (compatible with Python 3.13)
pandas>=2.1.0
numpy>=1.24.0

# Data Visualization
matplotlib>=3.5.0

# Excel File Handling
openpyxl>=3.0.0

# GUI Framework (Tkinter is included with Python)
# No separate installation needed for tkinter

# Additional utilities
pillow>=10.0.0

# Date and Time utilities
python-dateutil>=2.8.0
